+++ packages/react/src/theme/components/modal.ts
@@ -1,0 +1,14 @@
+import { modalAnatomy as parts } from "@chakra-ui/anatomy";
+import { createMultiStyleConfigHelpers } from "@chakra-ui/styled-system";
+
+const { definePartsStyle, defineMultiStyleConfig } = createMultiStyleConfigHelpers(parts.keys);
+
+const baseStyle = definePartsStyle({
+  overlay: {
+    bg: "blackAlpha.700",
+  },
+});
+
+export const modalTheme = defineMultiStyleConfig({
+  baseStyle,
+});